ongoing-race.service renamed to horse-race.service
