+++ src/app/pages/races/ongoing/ongoing.component.ts
@@ -9,7 +9,7 @@ import { FormsModule } from '@angular/forms';
 import { SharedModule } from '@app/shared/shared.module';
 import { PhaserCanvasComponent } from '@app/components/phaser-canvas/phaser-canvas.component';
 import { OngoingListUiComponent } from '@app/components/ongoing-list-ui/ongoing-list-ui.component';
-import { OngoingRaceService } from '@app/game/ongoing-race.service';
+import { OngoingRaceService } from '@app/game/horse-race.service';
 import { Observable, Subscription } from 'rxjs';
 import { skip, filter } from 'rxjs/operators';
 import { WindowContainerComponent } from '@app/components/base-components/window-container/window-container.component';
